working memory initial update
Added in working memory, size of which  is controlled by variable self.delta. Also made it so there are separate update values per partner, as the updating of UV happens all in one function/at one time, so one global UV doesn't work (even if it seems theoretically more plausible? unsure on that). Default update val is set to 0.02