State Value Works
State value no longer breaks for paired entries, but also doesn't really give an evaluation of any value anymore. This is possibly due to reduced entries - it works based on checking opponent moves only still, but even at max paired memory entries this is only four values. At least it works, considering it isn't being used - but if it does need to be used it will need to be reconsidered