Get a Graph Generated
We are now generating an Erdős-Rényi graph, and exporting it in a way agents can view (namely, a dict keyed by their ID number). We next need to add in functions to convert between graph and dict, for the purposes of updating the graph at the model level, when agents choose to change their connections.